+++ judgels-frontends/raphael/src/modules/api/sandalphon/problemBundle.ts
@@ -1,0 +1,16 @@
+export enum ItemType {
+  Statement = 'STATEMENT',
+  MultipleChoice = 'MULTIPLE_CHOICE',
+}
+
+export interface Item {
+  jid: string;
+  type: ItemType;
+  meta: string;
+  config: string;
+}
+
+export interface ProblemWorksheet {
+  reasonNotAllowedToSubmit?: string;
+  items: Item[];
+}
