+++ judgels-frontends/raphael/src/modules/api/sandalphon/submissionBundle.ts
@@ -1,0 +1,10 @@
+export interface ItemSubmission {
+  id: number;
+  jid: string;
+  containerJid: string;
+  problemJid: string;
+  itemJid: string;
+  answer: string;
+  userJid: string;
+  time: number;
+}
